Dependency on the scheme is removed
There's error message by Google Chrome:
> Mixed Content: The page at 'https://artspb.me/cv/' was loaded over HTTPS, but requested an insecure stylesheet 'http://fonts.googleapis.com/css?family=Lato:400,500,600,700|Open+Sans:400,500,600,700'. This request has been blocked; the content must be served over HTTPS.
Browser doesn't allow to use insecure resources on the secured page. In such a case additional user action is required which is inconvenient. The idea of the patch is to remove explicit indication of the scheme (http or https). It will allow browser to choose scheme depending on the context.